♻️ :: CI 에러 수정

+++ src/main/kotlin/com/dsmdeliverybackend/domain/post/domain/repository/vo/QueryOrderPostVO.kt
@@ -14,7 +14,7 @@ class QueryOrderPostVO @QueryProjection constructor(
 
     val profileImg: String,
 
-    val star: Int,
+    val star: Float,
 
     val productType: String
