+++ src/main/kotlin/com/dsmdeliverybackend/domain/post/domain/repository/vo/QueryRiderPostVO.kt
@@ -14,5 +14,5 @@ class QueryRiderPostVO @QueryProjection constructor(
 
     val profileImg: String,
 
-    val star: Int
+    val star: Float
 )
